requirements.txt
aiohttp~=3.8.3
lxml~=4.9.1
tenacity~=8.1.0
Brotli~=1.0.9
PyExecJS~=1.5.1
zhconv~=1.4.3
PyYaml~=6.0
EbookLib~=0.18
pillow-avif-plugin~=1.3.1
Pillow~=9.5.0
sqlmodel~=0.0.20
APScheduler~=3.10.1